pyproject.toml
[tool.poetry]
name = "cir-sniper"
version = "0.1.0"
homepage = "https://github.com/michaeltoohig/cir-sniper"
description = "A bespoke J2Store sale sniper.."
authors = ["Michael Toohig <michaeltoohig@gmail.com>"]
readme = "README.md"
license =  "MIT"

[tool.poetry.dependencies]
python = "^3.7,<4.0"
click = "^8.1.3"
python-dotenv = "^0.20.0"
tabulate = "^0.8.10"
termgraph = "^0.5.3"
black = {version="^22.3.0", optional=true}
isort = {version="^5.8.0", optional=true}
flake8 = {version="^3.9.2", optional=true}
flake8-docstrings = {version="^1.6.0", optional=true}
pytest = {version="^7.1.2", optional=true}
pytest-cov = {version="^2.12.0", optional=true}
bump2version = {version="^1.0.1", optional=true}
Flask = {version = "^2.2.2", optional = true}
tenacity = "^8.1.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "flake8-docstrings",
    "pytest-cov",
    "flask"
    ]

dev = ["bump2version"]

[tool.poetry.scripts]
cir-sniper = 'cir_sniper.cli.cli:cli'

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"